refactor: move list_voices to cli module
